% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ham_game_control.R
\name{ham_eval}
\alias{ham_eval}
\title{Evaluate a turn or range of turns in the Hamurabi game}
\usage{
ham_eval(prop_starved, av_acres, last_pop = NULL)
}
\arguments{
\item{prop_starved}{vector of at least one year's worth of proportion of people who starved}

\item{av_acres}{acres per person at end of game or turn}

\item{last_pop}{population at end of game or turn.  Only needed if you get state 3 "mediocre" and
need to get a message of how many people would like to see you assassinated}
}
\value{
A list with two elements: "eval" is an integer from 1 (worst) to 4 (best) evaluating performance; 
"mess" is a character string to be displayed explaining how well / badly you went.
}
\description{
Evaluate a turn or range of turns in the Hamurabi game
}
\examples{

ham_eval(0.01, 11)
ham_eval(c(0.01, 0.05, 0, 0, 0.1), 11, last_pop = 75)
ham_eval(c(0.01, 0.05, 0, 0, 0.1), 8)
ham_eval(c(0.01, 0.05, 0, 0.5, 0.1), 11)
}
