% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ham_game_control.R
\name{ham_inputs}
\alias{ham_inputs}
\title{A year of inputs}
\usage{
ham_inputs(last_year, y = 0)
}
\arguments{
\item{last_year}{list with two elements, state and changes, as generated by ham_year}

\item{y}{year}
}
\description{
Prompt and accept a year's worth of decisions from the Hamurabi player
}
