% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ham_game_control.R
\name{ham_report}
\alias{ham_report}
\title{Report the state of the city}
\usage{
ham_report(last_year, y = 0)
}
\arguments{
\item{last_year}{list with two elements, state and changes, as generated by ham_year}

\item{y}{year}
}
\description{
Report the state of the city for a given year
}
