% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ham_year.R
\name{ham_year}
\alias{ham_year}
\title{Modify the state of the city}
\usage{
ham_year(state)
}
\arguments{
\item{state}{list representing the current state of the city, with elements 
acres_owned, acres_planted, people, bushels_fed, bushels_store}
}
\value{
a new value of state
}
\description{
Modify the state of the city for the Hamurabi game
}
\details{
corresponds to the code starting at line 511 of the BASIC version
}
