% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ham_game_control.R
\name{hamurabi}
\alias{hamurabi}
\title{Play the Hamurabi game}
\usage{
hamurabi(years = 10)
}
\arguments{
\item{years}{length of term}
}
\value{
None.  This function is called for its side effect of playing a game.
}
\description{
Human interface to play the Hamurabi city management game
}
